Fixed external process execution
Fixed external process execution
